+++ insts/std/MMOVE/RVB.cpp
@@ -379,7 +379,7 @@ RVB::set_gains(float rvbtime)
    int    i, fpoint, nvals = 16;
    float  rescale, gain, dist, G1, temp = SR / MACH1;
    double adjust;
-   static float array[16] = {
+   static double array[16] = {
       0, .001, 10, .1, 25, .225, 35, .28, 50, .35, 65, .4, 85, .45, 95, .475
    };
